Precision improvements
* app_quark_eos_[...].py:
- Changed the p_center space and the max_step, as small stars are problematic to calculate precisely;

* perfect_fluid_star_family_tides.py, perfect_fluid_star_tides.py, star_family.py, star_structure.py:
- Changed the r_begin parameter to 1e-12, as the previous value is smaller than necessary and not explicit;
- Changed the atol parameter to 1e-21, to make sure the error is controlled by the rtol parameter;

* perfect_fluid_star_tides.py:
- Added the logic to plot the perturbation curves;

* star_structure.py:
- Changed the logic that checks if the pressure is outside the acceptable range, now using the correct value p_center instead of p_0.

The small step size used in the tidal equation solution is needed in the case of some EOSs (GM1). Besides that, changing the equation to solve H'/H instead of H is not a good idea, as the initial value for H'/H is l/r, and with a small initial r, H'/H is large. Also, using H'/H as the variable in the equation does not make the solver use larger step sizes.